Split the grid tools instantiations.
This commit splits the grid tools instantiation files in two: one file deals
with instantiations for DoFHandlers and Triangulations and the other deals with
instantiations of just Triangulations. Here is the timing information:

before this patch:
debug:
grid tools: 74s, 3.6 GB

release:
grid tools: 97s 3.9 GB

after this patch:
debug:
grid tools: 31s, 1.8 GB
grid tools dof handlers: 48s, 2.4 GB

release:
grid tools: 46s, 1.9 GB
grid tools dof handlers: 65s, 2.4 GB

This lowers the high water mark for memory usage with only a small impact on the
total compilation time.
